backends/fastapi/pyproject.toml
[project]
name = "the-forum-api"
version = "0.1.0"
description = "The Forum — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.34.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "python-dotenv>=1.0.1",
  "psycopg2-binary>=2.9.10",
  "sqlalchemy>=2.0.37",
  "google-api-python-client>=2.150.0",
  "google-auth-oauthlib>=1.2.0",
  "openai>=1.60.0",
  "beautifulsoup4>=4.12.0",
  "thefuzz[speedup]>=0.22.0",
  "python-dateutil>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
  "httpx>=0.28.0",
  "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
